refactor: rename employees to staff (Closes #101)
